add list-blueprint
- add mocked list-view using flask blueprint
- remove old list-view
- register blueprint in main
