+++ list-demo-api/app.yaml
@@ -3,7 +3,5 @@ api_version: 1
 threadsafe: true
 
 handlers:
-- url: /static
-  static_dir: static
 - url: /.*
   script: main.app
