+++ list-demo-api/main.py
@@ -1,14 +1,5 @@
-from flask import Flask, request
+from flask import Flask
+from views.list import list_bp
 
 app = Flask(__name__)
-
-
-@app.route('/api/list')
-def get_list():
-    return '{"list": ["milch", "eier", "brot"]}'
-
-
-@app.route('/api/list', methods=['POST'])
-def add_element_to_list():
-    element = request.form['element']
-    return element + ' was successfully added to the list!'
+app.register_blueprint(list_bp, url_prefix='/api')
